Work for getting Loans working better. This includes a generic "global" formatter for classes, which diddn't work completely before.
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@1084 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/tasks/RecordSetTask.java
@@ -371,7 +371,7 @@ public class RecordSetTask extends BaseTask
             {
                 RecordSetIFace srcRecordSet = (RecordSetIFace)srcObj;
                 RecordSetIFace dstRecordSet = (RecordSetIFace)dstObj;
-                if (srcRecordSet.getDbTableId() == dstRecordSet.getDbTableId())
+                if (srcRecordSet.getDbTableId().intValue() == dstRecordSet.getDbTableId().intValue())
                 {
                     int oldSize = dstRecordSet.getItems().size();
                     Vector<RecordSetItemIFace> dstList  = new Vector<RecordSetItemIFace>(dstRecordSet.getItems());
